requirements.txt
torch>=2.0.0
transformers>=4.37.0
qwen-vl-utils>=0.0.1
Pillow>=10.0.0
accelerate>=0.25.0
pandas>=2.0.0
pyarrow>=14.0.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
pydantic>=2.0.0

